requirements.txt
torch
transformers
datasets
accelerate
trl
peft
bitsandbytes
pyyaml
wandb